Make bash the default CMD in core
In ubuntu:jammy (our base image for core) the default CMD is 'bash'.
When we specify an entrypoint (to enforce the use of tini), the default
cmd is then set to 'null' which is not backward compatible for projects
that assume the base cmd of 'bash'. Overall the impact should be low,
since most other projects will either not need this or they will
override it.

Ref:
https://docs.docker.com/engine/reference/builder/#understand-how-cmd-and-entrypoint-interact

>If CMD is defined from the base image, setting ENTRYPOINT will reset
>CMD to an empty value. In this scenario, CMD must be defined in the
>current image to have a value.