Add all available java images (#70)
Adds:
- java:8
- java:8-jre
- java:17
- java:17-jre
- java:18
- java:18-jre
- java:19
- java:19-jre

Note: There is a pretty big workaround for java:19(-jre) due to an issue with ca-certificates-java not handling a post install script correctly. I can't seem to replicate the issue locally however. Sooner or later that workaround should be removed when the build is no longer failing without the workaround. See here:
https://github.com/get-bridge/docker-base-images/pull/70/files#diff-fa1a7d7ca3ea5b4b9377f24cc9ef046efb0b6e8f87f3f4fcbb26db52584d7b8bR17-R20